pi-thermostat-scaffold/requirements.txt
RPi.GPIO==0.7.1
pydantic==2.8.2
pyyaml==6.0.2
